More changes, trying to figure out MPI stall on 2048 procs on Ranger that only occurs on one of the early files created.  Only occurs when generally large files being generated -- e.g. 32x32x8 per CPU, but not 4x4x4 per CPU with 2048 procs in 16x16x8 ncpux?

+++ global.structs.h
@@ -9,14 +9,15 @@
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 
+// none of below blink integers should be nearly as large as 32-bit signed integer can handle
 struct blink {
-  long long int num;
+  int num;
   struct blink * np;
   // only used by cpu=0
-  long long int cpu; // which cpu
-  long long int i,j,k,col; // starting values for cpu=0
-  long long int ri,rj,rk,rcol; // reference values for first cpu in sequence of nodes for a single buffer
-  long long int end;
+  int cpu; // which cpu
+  int i,j,k,col; // starting values for cpu=0
+  int ri,rj,rk,rcol; // reference values for first cpu in sequence of nodes for a single buffer
+  int end;
 };
